pyproject.toml
[tool.poetry]
name = "microsoft-ai-challenge-day"
version = "0.1.0"
description = ""
authors = ["Masahiro Ogawa <masahiro.ogaw@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
azure-search-documents = "11.4.0"
openai = "^1.33.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
